requirements.txt
pyyaml
httpx~=0.28.1
starlette~=0.47.2
pydantic~=2.11.7
fastapi~=0.116-1
uvicorn~=0.35.0
sqlmodel~=0.0.24
alembic~=1.16.4